requirements.txt
discord.py[voice]>=2.3.2
audioop-lts; python_version >= "3.13"
python-dotenv>=1.0.0
yt-dlp>=2023.12.30
PyNaCl>=1.5.0
aiohttp>=3.9.1
psutil>=5.9.6
aiosqlite>=0.19.0
python-dateutil>=2.8.2
PyQt6>=6.6.1
pygame>=2.5.2
mutagen>=1.47.0
requests>=2.31.0
gTTS>=2.5.1